requirements.txt
reflex==0.6.4
openai==1.38.0
pytest==7.4.3
pytest-mock==3.12.0
postgrest-py==0.10.6
python-dotenv==1.0.0
pytest-asyncio==0.23.5
